server: reorganize dir structure
